refactor(racingCar): 객체의 책임과 역할 분리 - Validation, Parsing 클래스 생성
